+++ service/controller/control-plane/resource.go
@@ -14,6 +14,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/certificates"
 	etcdcertificates "github.com/giantswarm/prometheus-meta-operator/service/controller/resource/etcd-certificates"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/heartbeat"
+	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/heartbeatrouting"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/ingress"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/namespace"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/prometheus"
@@ -240,6 +241,21 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		}
 	}
 
+	var heartbeatRoutingResource resource.Interface
+	{
+		c := heartbeatrouting.Config{
+			Installation: config.Installation,
+			K8sClient:    config.K8sClient,
+			Logger:       config.Logger,
+			OpsgenieKey:  config.OpsgenieKey,
+		}
+
+		heartbeatRoutingResource, err = heartbeatrouting.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	resources := []resource.Interface{
 		namespaceResource,
 		tlsCertificatesResource,
@@ -254,6 +270,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		ingressResource,
 		promxyResource,
 		heartbeatResource,
+		heartbeatRoutingResource,
 	}
 
 	{