+++ service/controller/resource/heartbeatrouting/receiver/receiver.go
@@ -5,34 +5,34 @@ import (
 	"net/url"
 	"reflect"
 
-	"github.com/prometheus/alertmanager/config"
-	commoncfg "github.com/prometheus/common/config"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/microerror"
 
+	alertmanagerconfig "github.com/giantswarm/prometheus-meta-operator/pkg/alertmanager/config"
+	promcommonconfig "github.com/giantswarm/prometheus-meta-operator/pkg/prometheus/common/config"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
-func toReceiver(cluster metav1.Object, installation string, opsgenieKey string) (config.Receiver, error) {
+func toReceiver(cluster metav1.Object, installation string, opsgenieKey string) (alertmanagerconfig.Receiver, error) {
 	u, err := url.Parse(fmt.Sprintf("https://api.opsgenie.com/v2/heartbeats/%s/ping", key.HeartbeatName(cluster, installation)))
 	if err != nil {
-		return config.Receiver{}, microerror.Mask(err)
+		return alertmanagerconfig.Receiver{}, microerror.Mask(err)
 	}
 
-	r := config.Receiver{
+	r := alertmanagerconfig.Receiver{
 		Name: key.HeartbeatReceiverName(cluster, installation),
-		WebhookConfigs: []*config.WebhookConfig{
-			&config.WebhookConfig{
-				URL: &config.URL{
+		WebhookConfigs: []*alertmanagerconfig.WebhookConfig{
+			&alertmanagerconfig.WebhookConfig{
+				URL: &alertmanagerconfig.URL{
 					URL: u,
 				},
-				HTTPConfig: &commoncfg.HTTPClientConfig{
-					BasicAuth: &commoncfg.BasicAuth{
-						Password: commoncfg.Secret(opsgenieKey),
+				HTTPConfig: &promcommonconfig.HTTPClientConfig{
+					BasicAuth: &promcommonconfig.BasicAuth{
+						Password: promcommonconfig.Secret(opsgenieKey),
 					},
 				},
-				NotifierConfig: config.NotifierConfig{
+				NotifierConfig: alertmanagerconfig.NotifierConfig{
 					VSendResolved: false,
 				},
 			},
@@ -44,7 +44,7 @@ func toReceiver(cluster metav1.Object, installation string, opsgenieKey string) 
 
 // EnsureCreated ensure receiver exist in cfg.Receivers and is up to date. Returns true when changes have been made to cfg.
 // Return untouched cfg and false when no changes are made.
-func EnsureCreated(cfg config.Config, cluster metav1.Object, installation, opsgenieKey string) (config.Config, bool, error) {
+func EnsureCreated(cfg alertmanagerconfig.Config, cluster metav1.Object, installation, opsgenieKey string) (alertmanagerconfig.Config, bool, error) {
 	desired, err := toReceiver(cluster, installation, opsgenieKey)
 	if err != nil {
 		return cfg, false, microerror.Mask(err)
@@ -67,7 +67,7 @@ func EnsureCreated(cfg config.Config, cluster metav1.Object, installation, opsge
 
 // EnsureDeleted ensure receiver is removed from cfg.Receivers. Returns true when changes have been made to cfg.
 // Return untouched cfg and false when no changes are made.
-func EnsureDeleted(cfg config.Config, cluster metav1.Object, installation, opsgenieKey string) (config.Config, bool, error) {
+func EnsureDeleted(cfg alertmanagerconfig.Config, cluster metav1.Object, installation, opsgenieKey string) (alertmanagerconfig.Config, bool, error) {
 	desired, err := toReceiver(cluster, installation, opsgenieKey)
 	if err != nil {
 		return cfg, false, microerror.Mask(err)
@@ -83,7 +83,7 @@ func EnsureDeleted(cfg config.Config, cluster metav1.Object, installation, opsge
 	return cfg, false, nil
 }
 
-func getReceiver(cfg config.Config, receiver config.Receiver) (*config.Receiver, int) {
+func getReceiver(cfg alertmanagerconfig.Config, receiver alertmanagerconfig.Receiver) (*alertmanagerconfig.Receiver, int) {
 	for index, r := range cfg.Receivers {
 		if r.Name == receiver.Name {
 			return r, index